Add the game function for playing 5 rounds
The game function was added here which uses loop to count to number
of desired round and also used to keep records of who wins in the
total rounds.

Visiting the code will help explain better on the steps taken in this
game function as it seems to be the hardest.