Debug the game function
Noticed that the computerSelection random Number
generator was not properly called inin the game function
so this caused issue were the gmae function was just
recycling with one number for the computer play.

Also did some changes on the number of rounds from 5
rounds to who gets to score 5 points in the game wins.